moving redundant details views to auth app for both users
