+++ customer/urls.py
@@ -3,5 +3,4 @@ from . import views
 
 urlpatterns = [
     path("signup/", views.CustomerSignupView.as_view(), name="customer_signup"),
-    path("profile/", views.CustomerDetailsView.as_view(), name="customer_details"),
 ]
