+++ driver/urls.py
@@ -4,5 +4,4 @@ from . import views
 
 urlpatterns = [
     path("signup/", views.DriverSignupView.as_view(), name="driver_signup"),
-    path("profile/", views.DriverDetailsView.as_view(), name="driver_details"),
 ]
